requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-cors-headers==2.4.0
django-storages==1.7.1
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
whitenoise==4.1.2
